applied Amanieu's patch d24893d3dede which removes most warnings

+++ xreal/base/code/game/g_lua.c
@@ -230,7 +230,7 @@ void G_RunLuaFunction(const char *func, const char *sig, ...)
 		{
 			case 'f':
 				// float argument
-				lua_pushnumber(L, va_arg(vl, float));
+				lua_pushnumber(L, va_arg(vl, double));
 
 				break;
